// aiapp/frontend/src/pages/web-cloud/access/general/connections.types.ts
export interface ModemCredentials {
  mode: 'PPPoE' | 'DHCP';
  username: string;
  password: string;
  vlan?: number;
}

export interface ModemGuide {
  brand: string;
  model: string;
  url: string;
}